user cart and order finished

+++ app/src/main/java/com/example/seabuckcafe/ui/register/RegisterFragment.kt
@@ -92,7 +92,7 @@ class RegisterFragment : Fragment() {
                             firebaseUser.uid,
                             name,
                             email,
-                            binding.phoneNumberEditText.text.toString().toLong(),
+                            binding.phoneNumberEditText.text.toString(),
                             isUser = true)
 
                     Firestore().registerUser(this, userInfo)
